+++ src/components/Content.jsx
@@ -1,23 +1,78 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import {useSelector} from 'react-redux'
-import axios from 'axios';
 import ArrowBackIcon from '@material-ui/icons/ArrowBack';
 import PublicIcon from '@material-ui/icons/Public';
+import Typography from '@material-ui/core/Typography';
+import Skeleton from '@material-ui/lab/Skeleton';
 
 export const Content = ({
-    contentId,
     onClickBackButton,
 }) => {
-    const state = useSelector(state => state.contentState)
-
+    const state = useSelector(state => state)
+    console.log(state)
     return (
         <>
         {state.loading ? (
             <>
-                <div>loading...</div>
+            <div className="content">
+                <div className="content-header">
+                <ArrowBackIcon
+                className="post-icon-1"
+                onClick = {() => onClickBackButton()}
+                ></ArrowBackIcon>
+                <p>制度詳細</p>
+            </div>
+            <div className="content-detail">
+                <Typography key="skeleton-a" variant="h2">
+                    <Skeleton animation="wave"/>
+                </Typography>
+                <Typography key="skeleton-a" variant="h3">
+                    <Skeleton animation="wave"/>
+                </Typography>
+                <div className="content-container">
+                    <div>
+                        <Typography key="skeleton-a" variant="h4">
+                            <Skeleton animation="wave" width='30%'/>
+                        </Typography>
+                        <Typography key="skeleton-a" variant="a">
+                            <Skeleton animation="wave"/>
+                            <Skeleton animation="wave"/>
+                            <Skeleton animation="wave"/>
+                        </Typography>
+                    </div>
+                    <div>
+                        <Typography key="skeleton-a" variant="h4">
+                            <Skeleton animation="wave" width='30%'/>
+                        </Typography>
+                        <Typography key="skeleton-a" variant="a">
+                            <Skeleton animation="wave"/>
+                            <Skeleton animation="wave"/>
+                            <Skeleton animation="wave"/>
+                        </Typography>
+                    </div>
+                    <div>
+                        <Typography key="skeleton-a" variant="a">
+                            <Skeleton animation="wave"/>
+                            <Skeleton animation="wave"/>
+                            <Skeleton animation="wave"/>
+                        </Typography>
+                    </div>
+                    <div>
+                        <Typography key="skeleton-a" variant="h4">
+                            <Skeleton animation="wave" width='30%'/>
+                        </Typography>
+                        <Typography key="skeleton-a" variant="a">
+                            <Skeleton animation="wave"/>
+                            <Skeleton animation="wave"/>
+                            <Skeleton animation="wave"/>
+                        </Typography>
+                    </div>
+                </div>
+            </div>
+            </div>
             </>
             ) : (
-                <div className="content">
+            <div className="content">
                 <div className="content-header">
                 <ArrowBackIcon
                 className="post-icon-1"
@@ -44,10 +99,10 @@ export const Content = ({
                               <h4>連絡先</h4>
                               <a>{state.content.inquiry}</a>
                           </div>
+                          <div className="content-footer">
+                            <a href = {'https://seido-navi.mirasapo-plus.go.jp/supports/'+state.content.id}><PublicIcon className="post-icon-2"></PublicIcon><span>webpage</span></a>
+                          </div>
                   </div>
-                </div>
-                <div className="content-footer">
-                    <a href = {'https://seido-navi.mirasapo-plus.go.jp/supports/'+/*contentId*/1}><PublicIcon className="post-icon-2"></PublicIcon><span>webpage</span></a>
                 </div>
             </div>
             )
